requirements.txt
# requirements.txt
# Root Python dependencies for ARC Explainer (used by Docker build)
# Updated: 2026-01-31 - Created for Railway deployment
#
# Ethara's fork of litellm with custom patches
litellm @ git+https://github.com/Ethara-Ai/litellm.git@main
#
# ARCEngine core is vendored in external/ARCEngine

annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
arcengine==0.9.3
asynciolimiter>=1.2.0
boto3>=1.35.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
google-auth==2.48.0
google-genai==1.66.0
gymnasium==1.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio>=2.34.0
imageio-ffmpeg>=0.5.1
jiter==0.13.0
langchain-openai>=1.1.0
numpy==2.4.2
openai==2.24.0
pillow==12.1.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
scipy>=1.10.0
sniffio==1.3.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
